+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -1,7 +1,6 @@
 use std::collections::BTreeMap;
 
 use relay_base_schema::events::EventType;
-use relay_base_schema::spans::SpanStatus;
 use relay_common::time::UnixTimestamp;
 use relay_dynamic_config::{TagMapping, TransactionMetricsConfig};
 use relay_event_normalization::utils as normalize_utils;
@@ -24,13 +23,6 @@ pub mod types;
 /// Placeholder for transaction names in metrics that is used when SDKs are likely sending high
 /// cardinality data such as raw URLs.
 const PLACEHOLDER_UNPARAMETERIZED: &str = "<< unparameterized >>";
-
-/// Extract transaction status, defaulting to [`SpanStatus::Unknown`].
-///
-/// Must be consistent with `process_trace_context` in [`relay_event_normalization`].
-fn extract_transaction_status(trace_context: &TraceContext) -> SpanStatus {
-    *trace_context.status.value().unwrap_or(&SpanStatus::Unknown)
-}
 
 /// Extract HTTP method
 /// See <https://github.com/getsentry/snuba/blob/2e038c13a50735d58cc9397a29155ab5422a62e5/snuba/datasets/errors_processor.py#L64-L67>.
@@ -152,9 +144,11 @@ fn extract_universal_tags(event: &Event, config: &TransactionMetricsConfig) -> C
     tags.insert(CommonTag::Platform, platform.to_string());
 
     if let Some(trace_context) = event.context::<TraceContext>() {
-        let status = extract_transaction_status(trace_context);
-
-        tags.insert(CommonTag::TransactionStatus, status.to_string());
+        // We assume that the trace context status is automatically set to unknown inside of the
+        // light event normalization step.
+        if let Some(status) = trace_context.status.value() {
+            tags.insert(CommonTag::TransactionStatus, status.to_string());
+        }
 
         if let Some(op) = normalize_utils::extract_transaction_op(trace_context) {
             tags.insert(CommonTag::TransactionOp, op);
@@ -1018,7 +1012,11 @@ mod tests {
             "type": "transaction",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}}
+            "contexts": {
+                "trace": {
+                    "status": "ok"
+                }
+            }
         }
         "#;
 
@@ -1043,7 +1041,7 @@ mod tests {
         assert_eq!(
             extracted.project_metrics[0].tags,
             BTreeMap::from([
-                ("transaction.status".to_string(), "unknown".to_string()),
+                ("transaction.status".to_string(), "ok".to_string()),
                 ("platform".to_string(), "other".to_string())
             ])
         );
@@ -1051,12 +1049,17 @@ mod tests {
 
     #[test]
     fn test_span_tags() {
+        // Status is normalized upstream in the light normalization step.
         let json = r#"
         {
             "type": "transaction",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
+            "contexts": {
+                "trace": {
+                    "status": "ok"
+                }
+            },
             "spans": [
                 {
                     "description": "<OrganizationContext>",
@@ -1106,7 +1109,7 @@ mod tests {
         assert_eq!(
             extracted.project_metrics[0].tags,
             BTreeMap::from([
-                ("transaction.status".to_string(), "unknown".to_string()),
+                ("transaction.status".to_string(), "ok".to_string()),
                 ("platform".to_string(), "other".to_string()),
                 ("http.status_code".to_string(), "200".to_string())
             ])